refactor(sessions): modify how we check for sessions support
The test for whether a deployment supports sessions is to check if
`logicalSessionTimeoutMinutes` is set. For more complicated
deployments (e.g. not single) this value is not determined by a
single ismaster. As a result I've changed this to no longer be a
part of our `ServerCapabilities` detection, but read right off
the topology.

NODE-1188