fix(pool): ensure that lsid is sent in get requests to mongos
There is a strange case, when topology is mongos and readPreference
is not primary, where we nest our find queries in a "$query" attribute.
This was having the effect of not sending the lsid on find requests,
which caused further errors when getMores attempted to send an lsid.

Fixes NODE-1420