refactor: introduce knowledge of transactions to `executeOperation`
As this method is used in cursors, and does the required server
selection for operations in those cursors, it needs to know whether
we are in a transaction to provide some guarantees there. This is
part of unraveling the `applySession` megafunction we use at the
wire protocol level. Moving these concerns here is more logical
and decouples parts of the core driver.