fix(connect): prevent multiple callbacks in error scenarios
We treat `close`/`timeout`/`parseError`/`error` as error cases
during initial handshake, but some of these do not generate errors.
Additionally, in some cases multiple of these events may be emitted
which leaves us in a bad state if a subsequent `error` message is
emitted, since that will crash the node process.

NODE-2277