fix(connect): connect with family 0 instead of family 4
This is a fix for a hard-to-reproduce bug that occurs in
certain network setups on Windows 10. In  those cases,
attempting an IPv6 lookup with a shortname will somehow cause the
subsequent IPv4 lookup to also fail, even though it can succeed
independently. For some reason, this does NOT happen with family: 0.

This really should be fixed in either Node or in Windows itself,
but since we cannot create a stable reproducer, we will use this
fix short-term while we come up with a more stable solution.

Fixes NODE-1747
Fixes NODE-2143