refactor!: only store topology on MongoClient  (#2594)
A topology is now only stored on MongoClient and not on Db or
Collection. These classes now store a parent and access the 
topology through that parent. External topology accessors have
been removed, and an internal `getTopology` util has been
added. Methods that require a topology can now run into cases
where the topology is undefined because the client is closed; an 
error is thrown in these cases.

NODE-2850